+++ PortalDB/Models/QueryParams/Supply/EditSupplyItemQueryParams.cs
@@ -14,7 +14,7 @@ namespace PortalDB.Models.QueryParams.Supply
         [Required] public long? MeasurementUnitId { get; set; }
         [Required] public int? Quantity { get; set; }
         //[Required] public int? CurrentStock { get; set; }
-        [Required] public long? UnitCost { get; set; }
+        [Required] public decimal? UnitCost { get; set; }
         [Required] public int? ReorderPoint { get; set; }
         [Required] public long? StorageLocationId { get; set; }
         [Required] public long? VendorId { get; set; }
